splitting for the better
